# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library/books:
    get:
      operationId: listLibraryBooks
      tags: [library]
      summary: List the school's book catalog
      responses:
        "200":
          description: Books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryBook"
    post:
      operationId: createLibraryBook
      tags: [library]
      summary: Add a book to the catalog (admin/teacher)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryBookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBook"
  /library/books/{id}:
    patch:
      operationId: updateLibraryBook
      tags: [library]
      summary: Edit a catalog entry (admin/teacher)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryBookInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBook"
    delete:
      operationId: deleteLibraryBook
      tags: [library]
      summary: Remove a book from the catalog (admin/teacher)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /library/loans:
    get:
      operationId: listLibraryLoans
      tags: [library]
      summary: List loans (own children only for student/parent)
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryLoan"
    post:
      operationId: createLibraryLoan
      tags: [library]
      summary: Issue a book to a student (admin/teacher)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryLoanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryLoan"
  /library/loans/{id}/return:
    post:
      operationId: returnLibraryLoan
      tags: [library]
      summary: Mark a loan returned (admin/teacher)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryLoanReturnInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryLoan"

  # ── Transport ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LibraryBook:
      type: object
      required: [id, schoolId, title, author, totalCopies, availableCopies, createdAt]
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        title:
          type: string
        author:
          type: string
        isbn:
          type: ["string", "null"]
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        createdAt:
          type: string
          format: date-time
    LibraryBookInput:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: ["string", "null"]
        totalCopies:
          type: integer
    LibraryLoan:
      type: object
      required: [id, bookId, studentId, issuedAt, dueDate]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        studentId:
          type: integer
        issuedAt:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        returnedAt:
          type: ["string", "null"]
          format: date
        fineAmount:
          type: ["string", "null"]
        book:
          type: object
          properties:
            id:
              type: integer
            title:
              type: string
            author:
              type: string
        student:
          type: object
          properties:
            id:
              type: integer
            rollNo:
              type: string
            user:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
    LibraryLoanInput:
      type: object
      required: [bookId, studentId, dueDate]
      properties:
        bookId:
          type: integer
        studentId:
          type: integer
        dueDate:
          type: string
          format: date
    LibraryLoanReturnInput:
      type: object
      properties:
        fineAmount:
          type: number

    # ── Transport ─────────────────────────────────────────────────────────
...