# lib/api-zod/src/generated/types/libraryBook.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryBook {
  id: number;
  schoolId: number;
  title: string;
  author: string;
  /** @nullable */
  isbn?: string | null;
  totalCopies: number;
  availableCopies: number;
  createdAt: Date;
}