# lib/api-zod/src/generated/types/libraryBookInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryBookInput {
  title?: string;
  author?: string;
  /** @nullable */
  isbn?: string | null;
  totalCopies?: number;
}