# lib/api-zod/src/generated/types/libraryLoan.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * School Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { LibraryLoanBook } from './libraryLoanBook';
import type { LibraryLoanStudent } from './libraryLoanStudent';

export interface LibraryLoan {
  id: number;
  bookId: number;
  studentId: number;
  issuedAt: Date;
  dueDate: Date;
  /** @nullable */
  returnedAt?: Date | null;
  /** @nullable */
  fineAmount?: string | null;
  book?: LibraryLoanBook;
  student?: LibraryLoanStudent;
}